+++ tagless-final/src/main/java/com/github/jorgecastillo/kotlinandroid/tagless/algebras/business/model/SuperHero.kt
@@ -1,0 +1,3 @@
+package com.github.jorgecastillo.kotlinandroid.tagless.algebras.business.model
+
+data class SuperHero(val name: String)
