+++ tagless-final/src/main/java/com/github/jorgecastillo/kotlinandroid/tagless/runtime/instances/Instances.kt
@@ -5,7 +5,7 @@ import arrow.effects.IOHK
 import arrow.instance
 import com.github.jorgecastillo.kotlinandroid.tagless.algebras.persistence.DataSource
 import com.github.jorgecastillo.kotlinandroid.tagless.algebras.persistence.HeroesRepository
-import com.github.jorgecastillo.kotlinandroid.tagless.algebras.services.HeroesService
+import com.github.jorgecastillo.kotlinandroid.tagless.algebras.business.HeroesUseCases
 import com.github.jorgecastillo.kotlinandroid.tagless.algebras.ui.Navigation
 import com.github.jorgecastillo.kotlinandroid.tagless.algebras.ui.Presentation
 
@@ -18,7 +18,7 @@ interface IOPresentationInstance<F> : Presentation<F>
 interface IONavigationInstance<F> : Navigation<F>
 
 @instance(IO::class)
-interface IOHeroesServiceInstance<F> : HeroesService<F>
+interface IOHeroesUseCasesInstance<F> : HeroesUseCases<F>
 
 @instance(IO::class)
 interface IOHeroesRepositoryInstance<F> : HeroesRepository<F>
